+++ extensions/omni_schema/migrate/8_meta.sql
@@ -1,0 +1,3 @@
+-- create schema meta;
+/*{% include "../src/meta.sql" %}*/
+select instantiate_meta(schema => 'omni_schema');